rename: Group 에서 Plan 으로 변경
